Warn and continue when cached layer has no contents
